voiceforge/backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# AI/ML Models
torch>=2.1.0
torchaudio>=2.1.0
transformers==4.37.2
accelerate==0.26.1
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.4
scipy==1.12.0

# TTS Models
TTS==0.22.0  # Coqui TTS (includes XTTS)
openai-whisper==20231117  # Speech-to-text
faster-whisper==1.0.0  # Optimized Whisper

# Audio processing
pydub==0.25.1
audioread==3.0.1
noisereduce==3.0.2
demucs==4.0.1  # Voice isolation

# Utils
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.1
boto3==1.34.34  # S3 storage
celery==5.3.6  # Background tasks
flower==2.0.1  # Celery monitoring

# API Documentation
email-validator==2.1.0.post1
